Email Modifications
Email address is now required when creating a reservation.

Added a result Message (configurable via the admin interface) for when an email address is left blank.

Moved the email patron link to make it more visible.

Added a link to the email page so that you can click on the email and open in your default email app.